Refactor pda into a proper module
