+++ tgui/packages/tgui/interfaces/Pda/pda_screens/pda_janitor.tsx
@@ -1,5 +1,5 @@
-import { useBackend } from '../../backend';
-import { Box, LabeledList, Section } from '../../components';
+import { useBackend } from 'tgui/backend';
+import { Box, LabeledList, Section } from 'tgui/components';
 
 type Data = {
   janitor: {
