+++ tgui/packages/tgui/interfaces/Pda/pda_screens/pda_manifest.tsx
@@ -1,6 +1,7 @@
-import { useBackend } from '../../backend';
-import { Box } from '../../components';
-import { CrewManifestContent } from '../CrewManifest';
+import { useBackend } from 'tgui/backend';
+import { Box } from 'tgui/components';
+
+import { CrewManifestContent } from '../../CrewManifest';
 
 export const pda_manifest = (props) => {
   const { act, data } = useBackend();
