+++ tgui/packages/tgui/interfaces/Pda/pda_screens/pda_medical.tsx
@@ -1,6 +1,7 @@
-import { useBackend } from '../../backend';
-import { Box, Button, LabeledList, Section } from '../../components';
-import { GeneralRecord, MedicalRecord, RecordList } from './pda_types';
+import { useBackend } from 'tgui/backend';
+import { Box, Button, LabeledList, Section } from 'tgui/components';
+
+import { GeneralRecord, MedicalRecord, RecordList } from '../pda_types';
 
 type Data = {
   records: {
