+++ tgui/packages/tgui/interfaces/Pda/pda_screens/pda_notekeeper.tsx
@@ -1,6 +1,6 @@
 /* eslint react/no-danger: "off" */
-import { useBackend } from '../../backend';
-import { Box, Button, Section, Table } from '../../components';
+import { useBackend } from 'tgui/backend';
+import { Box, Button, Section, Table } from 'tgui/components';
 
 type Data = { note: string; notename: string };
