+++ tgui/packages/tgui/interfaces/Pda/pda_screens/pda_power.tsx
@@ -1,0 +1,5 @@
+import { PowerMonitorContent } from '../../PowerMonitor/PowerMonitorContent';
+
+export const pda_power = (props) => {
+  return <PowerMonitorContent />;
+};
