+++ tgui/packages/tgui/interfaces/Pda/pda_screens/pda_signaller.tsx
@@ -1,4 +1,4 @@
-import { SignalerContent } from '../Signaler';
+import { SignalerContent } from '../../Signaler';
 
 export const pda_signaller = (props) => {
   return <SignalerContent />;
