+++ tgui/packages/tgui/interfaces/Pda/pda_types.ts
@@ -1,4 +1,4 @@
-import { IconProps } from '../../components/Icon';
+import { IconProps } from 'tgui/components/Icon';
 /**
  * Gernal Record data
  */
