--- tgui/packages/tgui/interfaces/pda/pda_power.tsx
@@ -1,5 +1,0 @@
-import { PowerMonitorContent } from '../PowerMonitor/PowerMonitorContent';
-
-export const pda_power = (props) => {
-  return <PowerMonitorContent />;
-};
